Fix EC2 password environment variable not being passed to container
The PASSWORD environment variable was being written to .env file during
EC2 instance creation but was not being passed through to the sphinx-swarm
container in the docker-compose configuration.

Added PASSWORD=$PASSWORD to the environment section in second-brain-2.yml
so that custom passwords specified during swarm creation are properly
applied to the admin user.